improved analytics to track which page user is on
